Station: add nil check to #print_icons
We assume that we will always be passed a station that has routes.

But, as I found while painfully writing our feature tests, if we have
not seeded our test environment database, this is not a safe assumption.

In order to prevent the index.html.erb from crashing, if we have a nil
return "" so that the view render (albeit incorrectly).